+++ crates/module-system/module-implementations/sov-evm/src/authenticate.rs
@@ -1,22 +1,27 @@
-use borsh::BorshDeserialize;
+use std::marker::PhantomData;
+
+use borsh::{BorshDeserialize, BorshSerialize};
 use reth_primitives::TransactionSigned;
 use sov_address::{EthereumAddress, FromVmAddress};
 use sov_modules_api::capabilities::{
-    fatal_deserialization_error, AuthenticationOutput, AuthorizationData, FatalError,
-    TransactionAuthenticator, UniquenessData,
+    self, fatal_deserialization_error, AuthenticationOutput, AuthorizationData,
+    BatchFromUnregisteredSequencer, FatalError, TransactionAuthenticator, UniquenessData,
+    UnregisteredAuthenticationError,
 };
 use sov_modules_api::macros::config_value;
 use sov_modules_api::runtime::capabilities::AuthenticationError;
 use sov_modules_api::transaction::{
     AuthenticatedTransactionAndRawHash, AuthenticatedTransactionData, Credentials, PriorityFeeBips,
-    TxDetails,
+    TransactionWithoutCall, TxDetails,
 };
-use sov_modules_api::{Amount, FullyBakedTx, ProvableStateReader, RawTx, Spec};
+use sov_modules_api::{
+    Amount, DispatchCall, FullyBakedTx, ProvableStateReader, RawTx, Runtime, Spec,
+};
 use sov_rollup_interface::TxHash;
 use sov_state::User;
 
 use crate::conversions::RlpConversionError;
-use crate::{CallMessage, RlpEvmTransaction};
+use crate::{call, CallMessage, RlpEvmTransaction};
 
 /// Authenticates a raw evm transaction.
 /// # Security
@@ -93,14 +98,157 @@ pub fn decode_evm_tx(raw_tx: &[u8]) -> Result<(RlpEvmTransaction, TransactionSig
 
 /// Indicates that a runtime supports the `Ethereum` transaction authenticator
 /// and provides suitable methods for encoding and decoding Ethereum transactions.
-pub trait EthereumAuthenticator<S: Spec>: TransactionAuthenticator<S> {
+pub trait EthereumAuthenticator<S: Spec>: Runtime<S> {
     /// Add the Ethereum discriminant to a transaction the runtime.
-    fn add_ethereum_auth(tx: RawTx) -> <Self as TransactionAuthenticator<S>>::Input;
+    fn add_ethereum_auth(tx: RawTx) -> <Self::Auth as TransactionAuthenticator<S>>::Input;
 
     /// Encode a transaction with the Ethereum discriminant for the runtime.
     fn encode_with_ethereum_auth(tx: RawTx) -> FullyBakedTx {
-        <Self as TransactionAuthenticator<S>>::encode_authenticator_input(&Self::add_ethereum_auth(
-            tx,
-        ))
+        <Self::Auth as TransactionAuthenticator<S>>::encode_authenticator_input(
+            &Self::add_ethereum_auth(tx),
+        )
+    }
+}
+
+/// See [`TransactionAuthenticator::Input`].
+#[derive(std::fmt::Debug, Clone, BorshDeserialize, BorshSerialize)]
+pub enum EvmAuthenticatorInput<T = RawTx, U = RawTx> {
+    /// Authenticate using the `EVM` authenticator, which expects a standard EVM transaction
+    /// (i.e. an rlp-encoded payload signed using secp256k1 and hashed using keccak256).
+    Evm(T),
+    /// Authenticate using the standard `sov-module` authenticator, which uses the default
+    /// signature scheme and hashing algorithm defined in the rollup's [`Spec`].
+    Standard(U),
+}
+
+/// EVM-compatible transaction authenticator. See [`TransactionAuthenticator`].
+pub struct EvmAuthenticator<S, Rt>(PhantomData<(S, Rt)>);
+
+impl<S, Rt> TransactionAuthenticator<S> for EvmAuthenticator<S, Rt>
+where
+    S: Spec,
+    S::Address: FromVmAddress<EthereumAddress>,
+    Rt: Runtime<S> + DispatchCall<Spec = S>,
+{
+    type Decodable = EvmAuthenticatorInput<call::CallMessage, <Rt as DispatchCall>::Decodable>;
+    type Input = EvmAuthenticatorInput;
+    type Signature = EvmAuthenticatorInput<TransactionSigned, TransactionWithoutCall<S>>;
+
+    #[cfg(feature = "native")]
+    fn decode_serialized_tx(
+        tx: &FullyBakedTx,
+    ) -> Result<(Self::Decodable, Self::Signature), sov_modules_api::capabilities::FatalError> {
+        let auth_variant: EvmAuthenticatorInput = borsh::from_slice(&tx.data).map_err(|e| {
+            sov_modules_api::capabilities::FatalError::DeserializationFailed(e.to_string())
+        })?;
+
+        match auth_variant {
+            EvmAuthenticatorInput::Evm(raw_tx) => {
+                let (call, tx) = decode_evm_tx(&raw_tx.data)?;
+                Ok((
+                    EvmAuthenticatorInput::Evm(call::CallMessage { rlp: call }),
+                    EvmAuthenticatorInput::Evm(tx),
+                ))
+            }
+            EvmAuthenticatorInput::Standard(raw_tx) => {
+                let (call, tx) = capabilities::decode_sov_tx::<S, Rt>(&raw_tx.data)?;
+                Ok((
+                    EvmAuthenticatorInput::Standard(call),
+                    EvmAuthenticatorInput::Standard(tx),
+                ))
+            }
+        }
+    }
+
+    fn authenticate<Accessor: ProvableStateReader<User, Spec = S>>(
+        tx: &FullyBakedTx,
+        state: &mut Accessor,
+    ) -> Result<
+        capabilities::AuthenticationOutput<S, Self::Decodable>,
+        capabilities::AuthenticationError,
+    > {
+        let input: EvmAuthenticatorInput = borsh::from_slice(&tx.data).map_err(|e| {
+            sov_modules_api::capabilities::fatal_deserialization_error::<_, S, _>(
+                &tx.data, e, state,
+            )
+        })?;
+
+        match input {
+            EvmAuthenticatorInput::Evm(tx) => {
+                let (tx_and_raw_hash, auth_data, runtime_call) =
+                    authenticate::<_, _>(&tx.data, state)?;
+
+                Ok((
+                    tx_and_raw_hash,
+                    auth_data,
+                    EvmAuthenticatorInput::Evm(runtime_call),
+                ))
+            }
+            EvmAuthenticatorInput::Standard(tx) => {
+                let (tx_and_raw_hash, auth_data, runtime_call) =
+                    sov_modules_api::capabilities::authenticate::<_, S, Rt>(
+                        &tx.data,
+                        &Rt::CHAIN_HASH,
+                        state,
+                    )
+                    .unwrap();
+
+                Ok((
+                    tx_and_raw_hash,
+                    auth_data,
+                    EvmAuthenticatorInput::Standard(runtime_call),
+                ))
+            }
+        }
+    }
+
+    #[cfg(feature = "native")]
+    fn compute_tx_hash(
+        tx: &sov_modules_api::FullyBakedTx,
+    ) -> anyhow::Result<sov_modules_api::TxHash> {
+        let input: EvmAuthenticatorInput = borsh::from_slice(&tx.data)?;
+
+        match input {
+            EvmAuthenticatorInput::Evm(tx) => {
+                let (_rlp, tx) = decode_evm_tx(&tx.data)?;
+                Ok(TxHash::new(tx.hash().into()))
+            }
+            EvmAuthenticatorInput::Standard(tx) => Ok(capabilities::calculate_hash(
+                &tx.data,
+                &mut sov_modules_api::gas::UnlimitedGasMeter::<S>::default(),
+            )?),
+        }
+    }
+
+    fn authenticate_unregistered<Accessor: ProvableStateReader<User, Spec = S>>(
+        batch: &BatchFromUnregisteredSequencer,
+        state: &mut Accessor,
+    ) -> Result<
+        capabilities::AuthenticationOutput<S, Self::Decodable>,
+        capabilities::UnregisteredAuthenticationError,
+    > {
+        let (tx_and_raw_hash, auth_data, runtime_call) =
+            sov_modules_api::capabilities::RollupAuthenticator::<S, Rt>::authenticate_unregistered(
+                batch, state,
+            )?;
+
+        if Rt::allow_unregistered_tx(&runtime_call) {
+            Ok((
+                tx_and_raw_hash,
+                auth_data,
+                EvmAuthenticatorInput::Standard(runtime_call),
+            ))
+        } else {
+            Err(UnregisteredAuthenticationError::FatalError(
+                FatalError::Other(
+                    "The runtime call included in the transaction was invalid.".to_string(),
+                ),
+                tx_and_raw_hash.raw_tx_hash,
+            ))?
+        }
+    }
+
+    fn add_standard_auth(tx: RawTx) -> Self::Input {
+        EvmAuthenticatorInput::Standard(tx)
     }
 }