+++ crates/module-system/module-implementations/sov-evm/src/lib.rs
@@ -26,7 +26,9 @@ mod event;
 #[cfg(feature = "native")]
 mod helpers;
 
-pub use authenticate::{authenticate, decode_evm_tx, EthereumAuthenticator};
+pub use authenticate::{
+    authenticate, decode_evm_tx, EthereumAuthenticator, EvmAuthenticator, EvmAuthenticatorInput,
+};
 pub use reth_primitives::revm_primitives::SpecId;
 use reth_primitives::revm_primitives::{Address, BlockEnv, B256};
 pub use reth_primitives::TransactionSigned;