+++ crates/module-system/sov-modules-api/src/runtime/mod.rs
@@ -34,7 +34,6 @@ pub trait Runtime<S: Spec>:
     DispatchCall<Spec = S>
     + HasCapabilities<S>
     + HasKernel<S>
-    + TransactionAuthenticator<S, Decodable = <Self as DispatchCall>::Decodable>
     + Genesis<Spec = S, Config = Self::GenesisConfig>
     + TxHooks<Spec = S>
     + BlockHooks<Spec = S>
@@ -53,6 +52,16 @@ pub trait Runtime<S: Spec>:
     /// GenesisInput type.
     type GenesisInput: std::fmt::Debug + Clone + Send + Sync;
 
+    /// Responsible for authenticating transactions.
+    type Auth: TransactionAuthenticator<S>;
+
+    /// Decodes serialized call message.
+    fn decode_call(
+        serialized_message: &[u8],
+    ) -> Result<<Self as DispatchCall>::Decodable, io::Error> {
+        decode_borsh_serialized_message::<<Self as DispatchCall>::Decodable>(serialized_message)
+    }
+
     /// Default RPC methods and Axum router.
     fn endpoints(storage: crate::rest::ApiState<S>) -> NodeEndpoints;
 
@@ -62,12 +71,17 @@ pub trait Runtime<S: Spec>:
     /// Gets the operating mode of the runtime (Zk or Optimistic).
     fn operating_mode(genesis: &Self::GenesisConfig) -> OperatingMode;
 
-    /// Decodes serialized call message.
-    fn decode_call(
-        serialized_message: &[u8],
-    ) -> Result<<Self as DispatchCall>::Decodable, io::Error> {
-        decode_borsh_serialized_message::<<Self as DispatchCall>::Decodable>(serialized_message)
-    }
+    /// Wraps [`TransactionAuthenticator::Input`] into a call message.
+    fn wrap_call(
+        auth_data: <Self::Auth as TransactionAuthenticator<S>>::Decodable,
+    ) -> Self::Decodable;
+
+    /// Determines whether a transaction is allowed to be submitted on-chain by
+    /// someone other than a registered sequencer.
+    ///
+    /// This is a low level security mechanism. Your runtime SHOULD only allow
+    /// `sov_sequencer_registry::CallMessage::Register` transactions here.
+    fn allow_unregistered_tx(call: &Self::Decodable) -> bool;
 }
 
 #[cfg(feature = "native")]
@@ -87,16 +101,15 @@ pub fn decode_borsh_serialized_message<T: borsh::BorshDeserialize>(
     Ok(res)
 }
 
-#[cfg(not(feature = "native"))]
 /// This trait has to be implemented by a runtime in order to be used in `StfBlueprint`.
 ///
 /// The `TxHooks` implementation sets up a transaction context based on the height at which it is
 /// to be executed.
+#[cfg(not(feature = "native"))]
 pub trait Runtime<S: Spec>:
     DispatchCall<Spec = S>
     + HasCapabilities<S>
     + HasKernel<S>
-    + TransactionAuthenticator<S, Decodable = <Self as DispatchCall>::Decodable>
     + Genesis<Spec = S, Config = Self::GenesisConfig>
     + TxHooks<Spec = S>
     + BlockHooks<Spec = S>
@@ -111,8 +124,23 @@ pub trait Runtime<S: Spec>:
     /// `GenesisConfig` type.
     type GenesisConfig: Clone + Send + Sync;
 
+    /// Responsible for authenticating transactions.
+    type Auth: TransactionAuthenticator<S>;
+
     /// Gets the operating mode of the runtime (Zk or Optimistic).
     fn operating_mode(genesis: &Self::GenesisConfig) -> OperatingMode;
+
+    /// Wraps [`TransactionAuthenticator::Input`] into a call message.
+    fn wrap_call(
+        auth_data: <Self::Auth as TransactionAuthenticator<S>>::Decodable,
+    ) -> Self::Decodable;
+
+    /// Determines whether a transaction is allowed to be submitted on-chain by
+    /// someone other than a registered sequencer.
+    ///
+    /// This is a low level security mechanism. Your runtime SHOULD only allow
+    /// `sov_sequencer_registry::CallMessage::Register` transactions here.
+    fn allow_unregistered_tx(call: &Self::Decodable) -> bool;
 }
 
 /// The return type of [`Runtime::endpoints`].